+++ Harris Pharmacy/HarrisPharamacy/HarrisPharamacy.App/Startup.cs
@@ -21,11 +21,15 @@ v 1.0		Taylor Adam		2019-11-19			Added Headers
 
 #endregion copyright
 
+using System;
+using System.Linq;
+using System.Net.Http;
 using HarrisPharmacy.App.Areas.Identity;
 using HarrisPharmacy.Data;
 using HarrisPharmacy.Data.Interfaces;
 using HarrisPharmacy.Data.Services;
 using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Authorization;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
@@ -71,6 +75,21 @@ namespace HarrisPharmacy.App
             services.AddTransient<IFormService, FormService>();
             services.AddTransient<IAppointmentService, AppointmentService>();
             services.AddTransient<IPatientInfoService, PatientInfoService>();
+
+            // Server Side Blazor doesn't register HttpClient by default
+            if (services.All(x => x.ServiceType != typeof(HttpClient)))
+            {
+                // Setup HttpClient for server side in a client side compatible fashion
+                services.AddScoped<HttpClient>(s =>
+                {
+                    // Creating the URI helper needs to wait until the JS Runtime is initialized, so defer it.
+                    var uriHelper = s.GetRequiredService<NavigationManager>();
+                    return new HttpClient
+                    {
+                        BaseAddress = new Uri("https://localhost:44306/")
+                    };
+                });
+            }
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
